refactor: convert Sorter into SortService
