+++ src/province/province.dto.ts
@@ -2,7 +2,7 @@ import { IntersectionType, PartialType, PickType } from '@nestjs/mapped-types';
 import { IsNotEmpty, IsNumberString, Length } from 'class-validator';
 import { EqualsAny } from '@/common/decorator/EqualsAny';
 import { IsNotSymbol } from '@/common/decorator/IsNotSymbol';
-import { SortQuery } from '~/utils/helpers/sorter';
+import { SortQuery } from '@/sort/sort.dto';
 import { RegencySortQuery } from '../regency/regency.dto';
 
 export class Province {
